requirements.txt
betfairlightweight
boto3
matplotlib
numpy
orjson
pandas
pytest
python-dotenv
pytz
scikit-learn
shap
streamlit
tweepy
xgboost
python-telegram-bot>=20.0
fastapi
uvicorn
aiohttp
lxml
requests
tomli
